LPS-202515 Merge ScimGroupUtil and ScimUserUtil into the ScimUtil class to reuse methods between them more easily
